fixed variable type bug

+++ src/main/main.go
@@ -12,6 +12,6 @@ func main() {
 	flag.IntVar(&port, "Port", 8080, "Port the server listens to")
 
 	flag.Parse()
-	server.Run(port)
+	server.Run(uint16(port))
 
 }
